feat: predictor logging integration

+++ src/Open_MAGVIT2/modules/predictor/objectives/predictor_loss.py
@@ -12,6 +12,6 @@ def shift_ce_loss(
     codebook_size: int,
 ) -> torch.Tensor:
     return F.cross_entropy(
-        logits.reshape(-1, codebook_size),
+        logits.reshape(-1, codebook_size).float(),
         targets.reshape(-1),
     )
